更新演示版

+++ iphone/HospitalMobile/HospitalMobile/code/utility/LZUtility.m
@@ -42,7 +42,7 @@
     
     unsigned unitFlags = NSHourCalendarUnit|NSMinuteCalendarUnit;
     NSDateComponents * currentComp = [calendar components:unitFlags fromDate:dt];
-    int hour = [currentComp hour];//([currentComp hour]+ ([currentComp minute]>0?1:0))%24;
+    NSInteger hour = [currentComp hour];//([currentComp hour]+ ([currentComp minute]>0?1:0))%24;
     if (hour < 12)
     {
         return @"上午";
